Case studies & services
Added case study as an optional argument under project cards, for when I do a Medium write-up on a project that's a deep-dive into the challenges faced and successful resolutions. Also added a services section that is still a WIP, but will hopefully eventually detail the things that Meese Enterprises can offer prospective clients.